fix: fixed project structure
Related work item: GH-6
