+++ package.json
@@ -24,8 +24,6 @@
     "@mui/material": "^5.10.15",
     "formik": "^2.2.9",
     "i18next": "^22.0.6",
-    "i18next-browser-languagedetector": "^7.0.1",
-    "i18next-http-backend": "^2.1.0",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-i18next": "^12.0.0",
