+++ src/App.tsx
@@ -1,11 +1,6 @@
-import { customTheme } from './common/breakpoints';
+/*import { Route, Routes } from 'react-router-dom';*/
 import { Footer } from './templates/Footer/Footer';
-import { ThemeProvider } from '@mui/material/styles';
 
 export function App() {
-  return (
-    <ThemeProvider theme={customTheme}>
-      <Footer></Footer>
-    </ThemeProvider>
-  );
+  return <Footer></Footer>;
 }
