--- src/locales/Footer/en.json
@@ -1,7 +1,0 @@
-{
-    "common":{
-        "copyright":"Copyright © 2022",
-        "authors": "DSRM Team.",
-        "policy": "Policy terms"
-    }
-}
