--- src/locales/Footer/pl.json
@@ -1,7 +1,0 @@
-{
-    "common":{
-        "copyright":"Copyright © 2022",
-        "authors": "Zespół DSRM.",
-        "policy": "Policy terms"
-    }
-}
