+++ src/locales/en/footer.json
@@ -1,0 +1,5 @@
+{
+    "copyright":"Copyright © 2022",
+    "authors": "DSRM Team.",
+    "policy": "Policy terms"
+}
