+++ src/locales/pl/footer.json
@@ -1,0 +1,5 @@
+{
+    "copyright":"Copyright © 2022",
+    "authors": "Zespół DSRM.",
+    "policy": "Polityka prywatności"
+}
