--- src/react-i18next.d.ts
@@ -1,9 +1,0 @@
-import 'react-i18next';
-import en from './locales/Footer/en.json';
-
-declare module 'react-i18next' {
-  interface CustomTypeOptions {
-    defaultNS: 'common';
-    resources: typeof en;
-  }
-}
